+++ graphs/src/anemoi/graphs/describe.py
@@ -24,7 +24,7 @@ class GraphDescriptor:
 
     def __init__(self, path: Union[str, Path], **kwargs):
         self.path = path
-        self.graph = torch.load(self.path, weights_only=False)
+        self.graph = torch.load(self.path, weights_only=False, map_location="cpu")
 
     @property
     def total_size(self):