+++ graphs/src/anemoi/graphs/nodes/attributes/__init__.py
@@ -1,0 +1,28 @@
+# (C) Copyright 2024- Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+from .area_weights import PlanarAreaWeights
+from .area_weights import SphericalAreaWeights
+from .area_weights import UniformWeights
+from .boolean_op import BooleanAndMask
+from .boolean_op import BooleanNot
+from .boolean_op import BooleanOrMask
+from .masks import CutOutMask
+from .masks import NonmissingAnemoiDatasetVariable
+
+__all__ = [
+    "SphericalAreaWeights",
+    "PlanarAreaWeights",
+    "UniformWeights",
+    "CutOutMask",
+    "NonmissingAnemoiDatasetVariable",
+    "BooleanAndMask",
+    "BooleanNot",
+    "BooleanOrMask",
+]