+++ training/src/anemoi/training/diagnostics/mlflow/utils.py
@@ -14,6 +14,8 @@ from typing import Any
 
 import requests
 
+from anemoi.utils.remote import robust
+
 
 def health_check(tracking_uri: str) -> None:
     """Query the health endpoint of an MLflow server.
@@ -29,7 +31,11 @@ def health_check(tracking_uri: str) -> None:
     token = os.getenv("MLFLOW_TRACKING_TOKEN")
 
     headers = {"Authorization": f"Bearer {token}"}
-    response = requests.get(f"{tracking_uri}/health", headers=headers, timeout=60)
+    response = robust(requests.get, retry_after=30, maximum_tries=10)(
+        f"{tracking_uri}/health",
+        headers=headers,
+        timeout=60,
+    )
 
     if response.text == "OK":
         return