+++ training/src/anemoi/training/utils/usable_indices.py
@@ -14,13 +14,12 @@ import numpy as np
 
 
 def get_usable_indices(
-    missing_indices: set[int] | None,
+    missing_indices: set[int],
     series_length: int,
-    rollout: int,
-    multistep: int,
-    timeincrement: int = 1,
+    relative_indices: np.ndarray,
+    model_run_ids: np.ndarray | None = None,
 ) -> np.ndarray:
-    """Get the usable indices of a series whit missing indices.
+    """Get the usable indices of a series with missing indices.
 
     Parameters
     ----------
@@ -28,32 +27,28 @@ def get_usable_indices(
         Dataset to be used.
     series_length : int
         Length of the series.
-    rollout : int
-        Number of steps to roll out.
-    multistep : int
-        Number of previous indices to include as predictors.
-    timeincrement : int
-        Time increment, by default 1.
+    relative_indices: array[np.int64]
+        Array of relative indices requested at each index i.
+    model_run_ids: array[np.int64]
+        Array of integers of length series length that contains the id of a model run.
+        When training on analysis: None
 
     Returns
     -------
     usable_indices : np.array
         Array of usable indices.
     """
-    prev_invalid_dates = (multistep - 1) * timeincrement
-    next_invalid_dates = rollout * timeincrement
+    usable_indices = np.arange(series_length - max(relative_indices))
 
-    usable_indices = np.arange(series_length)  # set of all indices
-
-    if missing_indices is None:
-        missing_indices = set()
-
-    missing_indices |= {-1, series_length}  # to filter initial and final indices
+    # Avoid crossing model runs by selecting only relative indices with the same model run id
+    if model_run_ids is not None:
+        rel_run = usable_indices[None] + relative_indices[:, None]
+        include = (model_run_ids[rel_run] == model_run_ids[rel_run[0]]).all(axis=0)
+        usable_indices = usable_indices[include]
 
     # Missing indices
     for i in missing_indices:
-        usable_indices = usable_indices[
-            (usable_indices < i - next_invalid_dates) + (usable_indices > i + prev_invalid_dates)
-        ]
+        rel_missing = i - relative_indices  # indices which have their relative indices match the missing.
+        usable_indices = usable_indices[np.all(usable_indices != rel_missing[:, np.newaxis], axis=0)]
 
     return usable_indices